First Scenario working fine

+++ OptiboxSmart/app/src/main/java/com/example/optiboxsmart/clp_solver/Loader.java
@@ -198,9 +198,9 @@ public class Loader {
         return bestContainer;
     }
 
-    public List<double[]> solveToArray(){
+    public List<int[]> solveToArray(){
         Container container = solve();
-        List<double[]> l = new ArrayList<>();
+        List<int[]> l = new ArrayList<>();
         for (double[] d : container.toArray()){
 //            System.out.println("avant" + Arrays.toString(d));
             double [] pos = new double[]{d[0], d[1], d[2]};
@@ -208,13 +208,13 @@ public class Loader {
 //            System.out.println(dim[0]);
             // changement de repère (x,y,z) -> (y, z, x-X) (où X est la largeur du container
             double[] new_pos = new double[]{pos[1], pos[2], pos[0] - container.getDim(0)};
-            double[] pointsAG = new double[]{
-                    new_pos[0], //Ax
-                    new_pos[1], //Ay
-                    -(new_pos[2] + dim[0]), //Az
-                    new_pos[0] + dim[1], //Gx
-                    new_pos[1] + dim[2], //Gy
-                    -new_pos[2] //Gz
+            int[] pointsAG = new int[]{
+                    (int) new_pos[0], //Ax
+                    (int) new_pos[1], //Ay
+                    (int) (-(new_pos[2] + dim[0])), //Az
+                    (int) (new_pos[0] + dim[1]), //Gx
+                    (int) (new_pos[1] + dim[2]), //Gy
+                    (int) (-new_pos[2])//Gz
             };
 //            System.out.println("après" + Arrays.toString(pointsAG));
             l.add(pointsAG);
